+++ src/lua/lua_api.cpp
@@ -2940,6 +2940,18 @@ void load_api(LuaEnvironment& env)
 			stack.get_material(1)->set_vector3(stack.get_string_id_32(2), stack.get_vector3(3));
 			return 0;
 		});
+	env.add_module_function("Material", "set_vector4", [](lua_State* L)
+		{
+			LuaStack stack(L);
+			stack.get_material(1)->set_vector4(stack.get_string_id_32(2), stack.get_vector4(3));
+			return 0;
+		});
+	env.add_module_function("Material", "set_matrix4x4", [](lua_State* L)
+		{
+			LuaStack stack(L);
+			stack.get_material(1)->set_matrix4x4(stack.get_string_id_32(2), stack.get_matrix4x4(3));
+			return 0;
+		});
 
 	env.add_module_function("Gui", "move", [](lua_State* L)
 		{