+++ src/world/material.cpp
@@ -19,8 +19,8 @@ void Material::bind(ResourceManager& rm, ShaderManager& sm, u8 view, s32 depth) 
 	// Set samplers
 	for (u32 i = 0; i < _resource->num_textures; ++i)
 	{
-		const TextureData* td   = get_texture_data(_resource, i);
-		const TextureHandle* th = get_texture_handle(_resource, i, _data);
+		const TextureData* td   = texture_data(_resource, i);
+		const TextureHandle* th = texture_handle(_resource, i, _data);
 
 		const TextureResource* teximg = (TextureResource*)rm.get(RESOURCE_TYPE_TEXTURE, td->id);
 
@@ -39,7 +39,7 @@ void Material::bind(ResourceManager& rm, ShaderManager& sm, u8 view, s32 depth) 
 	// Set uniforms
 	for (u32 i = 0; i < _resource->num_uniforms; ++i)
 	{
-		const UniformHandle* uh = get_uniform_handle(_resource, i, _data);
+		const UniformHandle* uh = uniform_handle(_resource, i, _data);
 
 		bgfx::UniformHandle buh;
 		buh.idx = uh->uniform_handle;
@@ -51,20 +51,32 @@ void Material::bind(ResourceManager& rm, ShaderManager& sm, u8 view, s32 depth) 
 
 void Material::set_float(StringId32 name, f32 value)
 {
-	char* p = (char*)material_resource::get_uniform_handle_by_name(_resource, name, _data);
+	char* p = (char*)material_resource::uniform_handle_by_name(_resource, name, _data);
 	*(f32*)(p + sizeof(u32)) = value;
 }
 
 void Material::set_vector2(StringId32 name, const Vector2& value)
 {
-	char* p = (char*)material_resource::get_uniform_handle_by_name(_resource, name, _data);
+	char* p = (char*)material_resource::uniform_handle_by_name(_resource, name, _data);
 	*(Vector2*)(p + sizeof(u32)) = value;
 }
 
 void Material::set_vector3(StringId32 name, const Vector3& value)
 {
-	char* p = (char*)material_resource::get_uniform_handle_by_name(_resource, name, _data);
+	char* p = (char*)material_resource::uniform_handle_by_name(_resource, name, _data);
 	*(Vector3*)(p + sizeof(u32)) = value;
+}
+
+void Material::set_vector4(StringId32 name, const Vector4& value)
+{
+	char* p = (char*)material_resource::uniform_handle_by_name(_resource, name, _data);
+	*(Vector4*)(p + sizeof(u32)) = value;
+}
+
+void Material::set_matrix4x4(StringId32 name, const Matrix4x4& value)
+{
+	char* p = (char*)material_resource::uniform_handle_by_name(_resource, name, _data);
+	*(Matrix4x4*)(p + sizeof(u32)) = value;
 }
 
 } // namespace crown